+++ components/HeaderLayout/HeaderLayout.module.scss
@@ -1,0 +1,6 @@
+.outerWrapper {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  // padding: 1rem;
+}
