--- components/Tour/Tour.module.scss
@@ -1,16 +1,0 @@
-@use "../../styles/colors" as c;
-@use "../../styles/spacing" as s;
-
-.tourStartButton {
-  position: absolute;
-  left: 0;
-  top: 0;
-  background-color: c.$lighter-black;
-  border-bottom-right-radius: s.$border-radius;
-  padding: s.$chat-padding;
-  cursor: pointer;
-
-  &:active {
-    background-color: c.$bg-black;
-  }
-}
