+++ components/Tour/index.js
@@ -1,19 +1,12 @@
 import React, { useEffect } from "react";
-import s from "./Tour.module.scss";
 import JoyRide, { ACTIONS, EVENTS, STATUS } from "react-joyride";
 
-const Tour = ({ tourState, dispatch, refLines, setRefLines }) => {
+const Tour = ({ tourState, dispatch, refLines }) => {
   useEffect(() => {
     if (!localStorage.getItem("tour")) {
       dispatch({ type: "START" });
     }
   }, [dispatch]);
-
-  const handleStartTour = () => {
-    setRefLines([]);
-    localStorage.removeItem("tour");
-    dispatch({ type: "RESTART" });
-  };
 
   const callback = ({ action, index, type, status }) => {
     if (
@@ -65,9 +58,6 @@ const Tour = ({ tourState, dispatch, refLines, setRefLines }) => {
   };
   return (
     <>
-      <button className={s.tourStartButton} onClick={handleStartTour}>
-        Take a Tour
-      </button>
       <JoyRide
         {...tourState}
         showSkipButton={true}
