+++ components/VideoPlayer/VideoPlayer.module.scss
@@ -1,10 +1,11 @@
 @use '../../styles/colors' as c;
+@use '../../styles/spacing' as s;
 
 .mainSection {
   display: flex;
   flex: 1 1 0;
   // TODO: Make this less of a 'magic number' so that it always matches the height of the navbar above it
-  height: calc(100% - 35px);
+  height: calc(100% - s.$navbar-height);
 }
 
 .outerWrapper {
