+++ styles/Home.module.scss
@@ -27,6 +27,11 @@
   }
 }
 
+.tourStartButton {
+  background-color: inherit;
+  cursor: pointer;
+}
+
 .textWrapper {
   line-height: 200%;
 }
