+++ styles/_spacing.scss
@@ -1,3 +1,5 @@
+$navbar-height: 4.25rem;
+
 $text-box-height: 4.4rem;
 $border-radius: 14px;
 $gap: 0.6rem;
